refactor(bff): extract the management checks with an id/status parity pin
계층 정리 후속 1번. 행동 변화 없음 - 검사 4종과 확인 규칙을 자구 그대로 이동.

관리검사(인건비 이관, 입금 후 수익·부가세, 잔액 마이너스, 선입금 초과)의 계산은
BFF 가 소유하고 JVM 은 ID·상태의 형태만 @Pattern 으로 강제한다. 그 어휘가 갈리면
BFF 가 만든 검사 결과가 JVM Bean Validation 에서 400 으로 거부된다 - parity 표가
없던 마지막 큰 갭이었다.

  - cashflow-management-checks.mjs: 검사 4종 + 확인 매칭, 4800줄 라우트에서 추출.
    HTTP 도 Firestore 도 모른다. canonical 주차 재구성은 라우트의 builder 에 남는다.
  - cashflow-amounts.mjs / cashflow-range.mjs: 공용 순수 헬퍼 분리.
  - parity: BFF 표(ID 4종 + OK/WARNING/REVIEW_REQUIRED)와 JVM @Pattern 정규식을
    같은 리터럴로 고정. JVM 쪽은 record component 애노테이션을 리플렉션으로 읽어
    대조한다 - 어느 쪽을 고치든 상대 표가 깨진다.

라우트 파일 4814 -> 4670줄. 검사 동작 자체도 단위 테스트로 처음 고정
(미입력 WARNING / 0원 계획 REVIEW_REQUIRED / 기초잔액 누적).
